HttpClientとApiRequestを共通のprotocolとして分離する
